+++ .github/workflows/ci.yml
@@ -1,5 +1,5 @@
 ---
-name: Github CI
+name: CI
 on:
   pull_request:
     branches:
@@ -10,46 +10,6 @@ on:
       - master
       - beta
 jobs:
-  test:
-    runs-on: ubuntu-latest
-    steps:
-    - name: Checkout
-      uses: actions/checkout@v2
-      with:
-        fetch-depth: 0
-
-    - name: Use Node.js
-      uses: actions/setup-node@v2
-      with:
-        node-version: 16.x
-        cache: 'npm'
-
-    - name: Install
-      run: npm ci
-
-    - name: Lint commit messages
-      uses: wagoid/commitlint-github-action@v3
-
-    - name: ESLint
-      run: npm run lint
-
-    - name: Test
-      env:
-        SAUCE_USERNAME: ${{ secrets.SAUCE_USERNAME }}
-        SAUCE_ACCESS_KEY: ${{ secrets.SAUCE_ACCESS_KEY }}
-      run: |
-        echo "::add-matcher::.github/problem-matcher.json"
-        npm test
-
-    - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v2
-      with:
-        token: ${{ secrets.CODECOV_TOKEN }}
-
-    - name: Semantic release
-      run: npx semantic-release
-      if: success()
-      env:
-        HUSKY: 0
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
+  build:
+    uses: Neovici/cfg/.github/workflows/forge.yml@master
+    secrets: inherit
