+++ tsconfig.json
@@ -1,0 +1,14 @@
+{
+	"compilerOptions": {
+		"allowSyntheticDefaultImports": true,
+		"noEmit": true,
+		"module": "esnext",
+		"moduleResolution": "node",
+		"strict": true,
+		"target": "esnext",
+		"allowJs": true,
+		"paths": {
+			"@neovici/cosmoz-utils/*": ["./node_modules/@neovici/cosmoz-utils/dist/*"]
+		}
+	}
+}
